Change Discount Numeric Type to byte and check if Discount code has expired

+++ src/Data/TizianaTerenzi.Data.Models/DiscountCode.cs
@@ -8,7 +8,7 @@
     {
         public string Name { get; set; }
 
-        public double Discount { get; set; }
+        public byte Discount { get; set; }
 
         public DateTime? ExpiresOn { get; set; }
     }
